skills/caicai-url-to-markdown/requirements.txt
# Python dependencies for caicai-url-to-markdown
# Install with: pip install -r requirements.txt

# Core dependencies
requests>=2.28.0
readability-lxml>=0.8.1
markdownify>=0.11.0
beautifulsoup4>=4.11.0
